// src/shared/types/pagination.ts
export interface PaginationParams {
    page: number
    limit: number
  }
  
  export interface PaginationMeta {
    currentPage: number
    itemsPerPage: number
    totalItems: number
    totalPages: number
    hasNextPage: boolean
    hasPreviousPage: boolean
  }
  
  export interface PaginatedResponse<T> {
    data: T[]
    meta: PaginationMeta
  }